refactor(mcp): compile the embedded MCP server desktop-only
Move wealthfolio-mcp, rmcp, and axum to desktop-only target dependencies and cfg(desktop)-gate the server/middleware/audit_sink/lockfile/tests submodules, so the mobile (iOS/Android) Tauri binary no longer links the MCP transport stack. RunningServer moves into mod.rs so McpServerState stays lightweight and cross-platform, keeping the Tauri command signatures unchanged. Verified: desktop build + mcp tests pass, and cargo check --target aarch64-apple-ios compiles without rmcp/axum/wealthfolio-mcp.
