+++ apps/tauri/Cargo.toml
@@ -20,7 +20,6 @@ wealthfolio-device-sync = { workspace = true }
 wealthfolio-ai = { workspace = true }
 wealthfolio-spending = { workspace = true }
 wealthfolio-agent-tools = { workspace = true }
-wealthfolio-mcp = { workspace = true }
 
 # Tauri
 tauri = { version = "2.10", features = [] }
@@ -53,8 +52,6 @@ base64 = "0.22"
 urlencoding = "2.1.3"
 
 # Embedded MCP server
-axum = { version = "0.8", default-features = false, features = ["http1", "tokio", "json"] }
-rmcp = { workspace = true }
 subtle = "2"
 tokio-util = "0.7"
 
@@ -85,6 +82,11 @@ crate-type = ["staticlib", "cdylib", "rlib"]
 tauri-plugin-updater = "2"
 tauri-plugin-window-state = "2"
 tauri-plugin-single-instance = { version = "2", features = ["deep-link"] }
+
+# Embedded MCP server (desktop-only; mobile builds exclude it)
+axum = { version = "0.8", default-features = false, features = ["http1", "tokio", "json"] }
+rmcp = { workspace = true }
+wealthfolio-mcp = { workspace = true }
 
 [target.'cfg(any(target_os = "android", target_os = "ios"))'.dependencies]
 tauri-plugin-haptics = "2"
